Move CacheState into CachedValue
